requirements.txt
numpy==1.22.2
imutils==0.5.4
screeninfo==0.8.1
pywin32==305
cryptography==36.0.2
rsa==4.8
scapy==2.4.5
requests==2.25.1
keyboard==0.13.5
Pillow==8.4.0
pynput==1.7.6
opencv-python==4.7.0.68